tomb_arena.html
Essentially a working CRUD setup to build location dictionaries for a new zonemap.
Inspection will show that the update function has been commented out. This is because the update function itself did not really work. At the 
end of the day I am comfortable saying that if you really want to change an entry, just edit->trash->make changes->save
That will work out.
At this point, this may be the only CRUD we need to proceed. So, on to sifting the file apart into a server-client relationship.
You will see what I mean.