text_fiction
Here you find a generic file structure which is still buggy. I will work on it.
There are a few instructions. In general you are safe in changing things in pink. Also a definition "def" must be above the function call:
def a _function():
   if:
    this
  else:
    that

def some_other_function():
  if:
    a_function()
  else:
    some other stuff
The theme here is that if you can match letters to the keyboard you can start using PYTHON. Seriously, you can often cut & paste 
and be just fine. And if you keep you file in just one long file it really is easier. Write a story, plot some choices, invent a whole world 
and use this simple software.
Just one more thing: remember it is only a game, a game written in PYTHON.