Four threaded Fear Thee
Not working, I have a revision which allows me to create a new character. I can stand in Pinklemire Street. The error when I 'move' is not unusual. On the fly, I am reducing the created files to wear and hold and also the weapons; short sword, dagger and wand.